+++ backend/src/data/mockFile.ts
@@ -1,4 +1,4 @@
-import { MulterFile } from "../users";
+import { MulterFile } from "../types/users";
 
 const mockFile: MulterFile = {
   fieldname: "image",
